Files converted to Typescript from Javascript. Logic separated out into different components with closely related functions and variables moved closer to the relevant components.
Tracing has been restructured for readability and separation of concerns. The package now first treats tracing/index as the entrypoint.

Pages have been renamed to include _ where page routes are not meant to exist.
