+++ .eslintignore
@@ -5,5 +5,4 @@ README.md
 *jest.config*
 package-lock.json
 package.json
-server.js
 pages/functions/tooltip.js
