+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "nodemon --require ./tracing.js ./node_modules/.bin/next dev",
+    "dev": "nodemon --require ./tracing/index.ts ./node_modules/.bin/next dev",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
@@ -20,12 +20,14 @@
     "@mui/icons-material": "^5.11.11",
     "@mui/material": "^5.11.15",
     "@observablehq/plot": "^0.6.5",
+    "@opentelemetry/api": "^1.4.1",
     "@opentelemetry/exporter-trace-otlp-http": "^0.37.0",
     "@opentelemetry/instrumentation": "^0.37.0",
     "@opentelemetry/instrumentation-http": "^0.36.1",
     "@opentelemetry/instrumentation-mongoose": "^0.32.1",
     "@opentelemetry/instrumentation-pg": "^0.35.0",
     "@opentelemetry/sdk-trace-node": "^1.11.0",
+    "@types/express": "^4.17.17",
     "@types/node": "^18.15.11",
     "@types/react": "^18.0.31",
     "@types/react-dom": "^18.0.11",
@@ -45,7 +47,8 @@
     "react-chartjs-2": "^5.2.0",
     "react-dom": "18.2.0",
     "socket.io": "^4.6.1",
-    "socket.io-client": "^4.6.1"
+    "socket.io-client": "^4.6.1",
+    "ts-node": "^10.9.1"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
