+++ pages/api/demoPg.ts
@@ -1,0 +1,8 @@
+import type { NextApiRequest, NextApiResponse } from 'next';
+
+const { pool } = require('../../tracing/databaseModels');
+
+export default async function handler(req: NextApiRequest, res: NextApiResponse) {
+  const users = await pool.query('SELECT * FROM users');
+  res.status(200).json(users.rows);
+}
