+++ styles/EndpointsTable.module.css
@@ -1,0 +1,7 @@
+.detailList {
+  background-color: #212529;
+  height: 85%;
+  border: solid 1px #212529;
+  /* border-radius: 1rem; */
+  /* overflow-y: scroll; */
+}
