--- styles/MainWaterfall.module.css
@@ -1,9 +1,0 @@
-.chart {
-  width: 100%;
-  height: 15%;
-}
-
-.highlight {
-  stroke: white;
-  stroke-width: 4;
-}
