+++ styles/Timeline.module.css
@@ -1,0 +1,9 @@
+.chart {
+  width: 100%;
+  height: 15%;
+}
+
+.highlight {
+  stroke: white;
+  stroke-width: 4;
+}
