--- styles/theme.ts
@@ -1,20 +1,0 @@
-import { ThemeOptions } from '@mui/material/styles';
-
-export const themeOptions: any = {
-  palette: {
-    mode: 'dark',
-    primary: {
-      main: '#212529',
-    },
-    secondary: {
-      main: '#212529',
-    },
-    background: {
-      default: '#212529',
-      paper: '#212529',
-    },
-    success: {
-      main: '#212529',
-    },
-  },
-};
