+++ tracing/index.ts
@@ -1,0 +1,6 @@
+const provider = require('./instrumentation');
+
+provider.register();
+
+require('./tracingServer');
+require('./mongooseSetup');
